Created a repository package and moved EmployeeRepository to it. Moved the subclasses to the model package
