+++ src/oops/SOLID/singleResponsibilityPrinciple/before/model/FullTimeEmployee.java
@@ -1,4 +1,4 @@
-package oops.SOLID.singleResponsibilityPrinciple.before;
+package oops.SOLID.singleResponsibilityPrinciple.before.model;
 
 import oops.SOLID.singleResponsibilityPrinciple.before.model.Employee;
